requirements.txt
pywebview>=4.0
Pillow>=9.0.0
keyboard>=0.13.5
pystray>=0.19.0
